third_party/blink/web_tests/external/wpt/css/css-values/calc-serialization-expected.txt

This is a testharness.js-based test.
FAIL calc() serialization assert_equals: Serialization of calc(10px + 1vmin + 10%) expected "calc(10% + 10px + 1vmin)" but got "calc(10px + 1vmin + 10%)"
Harness: the test ran to completion.
